refactor: Move SDK mint tests to their own files
