+++ apps/sdk-e2e/src/integration/transaction-kinetic-sdk-e2e.spec.ts
@@ -158,16 +158,6 @@ describe('KineticSdk (e2e)', () => {
     }
   })
 
-  it('should create an account using a mint', async () => {
-    const account = Keypair.random()
-    const tx = await sdk.createAccount({ owner: account, mint: usdMint })
-    expect(tx).not.toBeNull()
-    expect(tx.mint).toBe(usdMint)
-    const { signature, errors } = tx
-    expect(typeof signature).toBe('string')
-    expect(errors).toEqual([])
-  })
-
   it('should make a transfer using a mint', async () => {
     const tx = await sdk.makeTransfer({
       amount: '1',
@@ -183,10 +173,4 @@ describe('KineticSdk (e2e)', () => {
     expect(Number(amount)).toBe(1)
     expect(source).toBe(aliceKeypair.publicKey)
   })
-
-  it('should request an airdrop using a mint', async () => {
-    const tx = await sdk.requestAirdrop({ account: aliceKeypair.publicKey, amount: '20', mint: usdMint })
-    expect(tx).not.toBeNull()
-    expect(typeof tx.signature).toBe('string')
-  }, 30000)
 })
